# asm/amd64: avoids allocation in getting memory location