```
$ benchstat old.txt new.txt
goos: darwin
goarch: amd64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
cpu: VirtualApple @ 2.50GHz
                                    │   old.txt   │              new.txt              │
                                    │   sec/op    │   sec/op     vs base              │
Compilation/with_extern_cache-10      263.7µ ± 3%   251.5µ ± 3%  -4.63% (p=0.001 n=7)
Compilation/without_extern_cache-10   4.234m ± 0%   4.225m ± 1%       ~ (p=0.165 n=7)
geomean                               1.057m        1.031m       -2.45%

                                    │   old.txt    │              new.txt               │
                                    │     B/op     │     B/op      vs base              │
Compilation/with_extern_cache-10      53.56Ki ± 0%   53.56Ki ± 0%       ~ (p=0.872 n=7)
Compilation/without_extern_cache-10   1.198Mi ± 0%   1.195Mi ± 0%  -0.26% (p=0.001 n=7)
geomean                               256.3Ki        256.0Ki       -0.13%

                                    │   old.txt   │              new.txt               │
                                    │  allocs/op  │  allocs/op   vs base               │
Compilation/with_extern_cache-10       982.0 ± 0%    981.0 ± 0%        ~ (p=0.070 n=7)
Compilation/without_extern_cache-10   15.00k ± 0%   12.33k ± 0%  -17.81% (p=0.001 n=7)
geomean                               3.838k        3.477k        -9.39%

```